Module fee::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

GenesisConfig
Can be used to configure the genesis state of this pallet.

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageFeePoolAccountId

_GeneratedPrefixForStorageIssueFee

_GeneratedPrefixForStorageIssueGriefingCollateral

_GeneratedPrefixForStorageMaintainerAccountId

_GeneratedPrefixForStorageMaintainerRewards

_GeneratedPrefixForStorageNominationRewards

_GeneratedPrefixForStoragePremiumRedeemFee

_GeneratedPrefixForStoragePunishmentFee

_GeneratedPrefixForStorageRedeemFee

_GeneratedPrefixForStorageRefundFee

_GeneratedPrefixForStorageReplaceGriefingCollateral

_GeneratedPrefixForStorageVaultRewards

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Traits

Config
Configuration

Type Definitions

FeePoolAccountId
AccountId of the fee pool.

IssueFee
Issue

IssueGriefingCollateral
Default griefing collateral (e.g. DOT/KSM) as a percentage of the locked collateral of a Vault a user has to lock to issue tokens.

MaintainerAccountId
AccountId of the parachain maintainer.

MaintainerRewards
Percentage of fees allocated for development.

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

NominationRewards
Percentage of fees generated by nominated collateral that is given to nominators.

PremiumRedeemFee
Vault Registry

PunishmentFee
Fee that a Vault has to pay if it fails to execute redeem or replace requests (for redeem, on top of the slashed value of the request). The fee is paid in collateral based on the token amount at the current exchange rate.

RedeemFee
Redeem

RefundFee
Refund

ReplaceGriefingCollateral
Replace

VaultRewards
Parachain Fee Pool Distribution