Trait fee::pallet::Config

pub trait Config: Config + Config + Config<UnsignedFixedPoint = <Self as Config>::UnsignedFixedPoint, SignedFixedPoint = <Self as Config>::SignedFixedPoint> {
    type FeePalletId: Get<PalletId>;
    type WeightInfo: WeightInfo;
    type SignedFixedPoint: FixedPointNumber<Inner = <Self as Config>::SignedInner> + Encode + EncodeLike + Decode;
    type SignedInner: Debug + CheckedDiv + TryFrom<<Self as Config>::Balance> + TryInto<<Self as Config>::Balance> + MaybeSerializeDeserialize;
    type UnsignedFixedPoint: FixedPointNumber<Inner = <Self as Config>::UnsignedInner> + Encode + EncodeLike + Decode + MaybeSerializeDeserialize + TypeInfo;
    type UnsignedInner: Debug + One + CheckedMul + CheckedDiv + FixedPointOperand + AtLeast32BitUnsigned + Default + Encode + EncodeLike + Decode + MaybeSerializeDeserialize + TypeInfo;
    type VaultRewards: RewardsApi<(), VaultId<<Self as Config>::AccountId, CurrencyId<Self>>, <Self as Config>::Balance, CurrencyId = CurrencyId<Self>>;
    type VaultStaking: StakingApi<VaultId<<Self as Config>::AccountId, CurrencyId<Self>>, Self::Index, <Self as Config>::Balance> + RewardsApi<(Option<Self::Index>, VaultId<<Self as Config>::AccountId, CurrencyId<Self>>), Self::AccountId, <Self as Config>::Balance, CurrencyId = CurrencyId<Self>>;
    type OnSweep: OnSweep<Self::AccountId, Amount<Self>>;
    type MaxExpectedValue: Get<<Self as Config>::UnsignedFixedPoint>;
}

Configuration

The pallet’s configuration trait.

Required Associated Types

type FeePalletId: Get<PalletId>

The fee module id, used for deriving its sovereign account ID.

type WeightInfo: WeightInfo

Weight information for the extrinsics in this module.

type SignedFixedPoint: FixedPointNumber<Inner = <Self as Config>::SignedInner> + Encode + EncodeLike + Decode

Signed fixed point type.

type SignedInner: Debug + CheckedDiv + TryFrom<<Self as Config>::Balance> + TryInto<<Self as Config>::Balance> + MaybeSerializeDeserialize

The Inner type of the SignedFixedPoint.

type UnsignedFixedPoint: FixedPointNumber<Inner = <Self as Config>::UnsignedInner> + Encode + EncodeLike + Decode + MaybeSerializeDeserialize + TypeInfo

Unsigned fixed point type.

type UnsignedInner: Debug + One + CheckedMul + CheckedDiv + FixedPointOperand + AtLeast32BitUnsigned + Default + Encode + EncodeLike + Decode + MaybeSerializeDeserialize + TypeInfo

The Inner type of the UnsignedFixedPoint.

type VaultRewards: RewardsApi<(), VaultId<<Self as Config>::AccountId, CurrencyId<Self>>, <Self as Config>::Balance, CurrencyId = CurrencyId<Self>>

Vault reward pool.

type VaultStaking: StakingApi<VaultId<<Self as Config>::AccountId, CurrencyId<Self>>, Self::Index, <Self as Config>::Balance> + RewardsApi<(Option<Self::Index>, VaultId<<Self as Config>::AccountId, CurrencyId<Self>>), Self::AccountId, <Self as Config>::Balance, CurrencyId = CurrencyId<Self>>

Vault staking pool.

type OnSweep: OnSweep<Self::AccountId, Amount<Self>>

Handler to transfer undistributed rewards.

type MaxExpectedValue: Get<<Self as Config>::UnsignedFixedPoint>

Maximum expected value to set the storage fields to.